Day 4 FP20 Analytics January 2025 Challenge
Did the background and explanations, tweaked some buttons and finished the menu for the map. Next is to try to put more information on the available spaces using markers especially centered on arrivals by month and maybe a different visualization for the "Delays by carrier name". It's somewhere between 75% and 80% done.